# docs/swagger.yaml
basePath: /
definitions:
...
  doc.DNS:
    properties:
      active:
        type: boolean
      base_domain:
        type: string
      createdat:
        type: string
      id:
        type: integer
      name:
        type: string
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      project_id:
        type: string
      provider:
        type: string
      region:
        type: string
      tls:
        type: string
      zone_id:
        type: string
    type: object
...
  doc.Organization:
    properties:
      createdat:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      image:
        type: string
      members:
        items:
          $ref: '#/definitions/doc.OrganizationMembers'
        type: array
      name:
        type: string
      organization_plan:
        $ref: '#/definitions/doc.OrganizationPlan'
      organization_plan_id:
        type: integer
      plugins:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
    type: object
  doc.OrganizationMembers:
    properties:
      createdat:
        type: string
      id:
        type: integer
      organization:
        $ref: '#/definitions/doc.Organization'
      organization_id:
        type: integer
      user:
        $ref: '#/definitions/doc.User'
      user_id:
        type: integer
      user_role:
        $ref: '#/definitions/doc.UserRoleEnum'
    type: object
  doc.OrganizationPlan:
    properties:
      active:
        type: boolean
      attributes:
        type: string
      cluster:
        type: integer
      cores:
        type: integer
      createdat:
        type: string
      id:
        type: integer
      memory:
        type: integer
      name:
        type: string
      no_of_user:
        type: integer
      price:
        type: integer
      weight:
        type: integer
    type: object
...
  doc.Plugin:
    properties:
      active:
        type: boolean
      add_ons:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
      attributes:
        type: string
      createdat:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      is_add_on:
        type: boolean
      min_cpu:
        type: integer
      min_memory:
        type: integer
      name:
        type: string
      plugin_category:
        items:
          $ref: '#/definitions/doc.PluginCategory'
        type: array
      service_detail: {}
      source_url:
        type: string
      support_ci:
        type: boolean
    type: object
  doc.PluginCategory:
    properties:
      createdat:
        type: string
      description:
        type: string
      id:
        type: integer
      is_add_on:
        type: boolean
      name:
        type: string
      plugins:
        items:
          $ref: '#/definitions/doc.Plugin'
        type: array
    type: object
...
  doc.User:
    properties:
      active:
        type: boolean
      address_updated:
        type: boolean
      company:
        type: string
      createdat:
        type: string
      designation:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      image:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      quotas:
        $ref: '#/definitions/models.Quotas'
      reference:
        type: string
      used_demo:
        type: boolean
    type: object
...
  doc.UserRoleEnum:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Admin
    - Members
...
  models.Quotas:
    properties:
      user_organization:
        type: integer
      user_project:
        type: integer
    type: object
...
host: api.test.01cloud.dev
info:
  contact:
    email: info@01cloud.com
    name: API Support
  description: 01cloud-api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 01cloud-api
  version: 1.0.1
paths:
...
  /admin/dns:
    get:
      consumes:
      - application/json
      description: Get list of dns by organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.DNS'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Dnss by organization
      tags:
      - Dns
...
  /dns:
    get:
      consumes:
      - application/json
      description: Get list of dns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/doc.DNS'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get Dns list
      tags:
      - Dns
    post:
      consumes:
      - application/json
      description: Create a new dns with the input payload
      parameters:
      - description: Create Dns
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DNS'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/doc.DNS'
      security:
      - ApiKeyAuth: []
      summary: Create a new dns
      tags:
      - Dns
  /dns/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Dns with the input payload
      parameters:
      - description: Dns id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/doc.DNS'
      security:
      - ApiKeyAuth: []
      summary: Delete a Dns
      tags:
      - Dns
    get:
      consumes:
      - application/json
      description: Get Dns by id from token
      parameters:
      - description: Dns id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.DNS'
      security:
      - ApiKeyAuth: []
      summary: Get Dns by id
      tags:
      - Dns
    put:
      consumes:
      - application/json
      description: Update a Dns with the input payload
      parameters:
      - description: Dns id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Dns
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DNS'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.DNS'
      security:
      - ApiKeyAuth: []
      summary: Update a Dns
      tags:
      - Dns
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"